@@ -102,8 +102,8 @@
                      <version>${record-builder-version}</version>
                   </path>
                </annotationProcessorPaths>
-               <source>21</source>
-               <target>21</target>
+               <source>25</source>
+               <target>25</target>
                <compilerArgs>--enable-preview</compilerArgs>
             </configuration>
          </plugin>
